feat: Исправлена работа вывода первых N фильмов и нахождения пересечений в друзьях

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,16 +2,18 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundAnythingException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
 public class FilmService {
 
-    private FilmStorage storage;
+    private final FilmStorage storage;
 
     @Autowired
     public FilmService(FilmStorage filmStorage) {
@@ -19,22 +21,29 @@ public class FilmService {
     }
 
     public void addLike(Long filmId, Long userId) {
-        storage.findById(filmId).addLike(userId);
+        if (!Objects.isNull(filmId) && !Objects.isNull(userId) && filmId > 0 && userId > 0) {
+            storage.findById(filmId).addLike(userId);
+        } else {
+            throw new NotFoundAnythingException("Номер пользователя или фильма не может быть < 0 или null");
+        }
     }
 
     public void deleteLike(Long filmId, Long userId) {
-        storage.findById(filmId).deleteLike(userId);
+        if (!Objects.isNull(filmId) && !Objects.isNull(userId) && filmId > 0 && userId > 0) {
+            storage.findById(filmId).deleteLike(userId);
+        } else {
+            throw new NotFoundAnythingException("Номер пользователя или фильма не может быть < 0 или null");
+        }
     }
 
     public List<Film> findfirstNByLikes(Integer size) {
         return storage.findAll().stream()
-                .sorted((p0, p1) -> compare(p0, p1))
+                .sorted(this::compare)
                 .limit(size)
                 .collect(Collectors.toList());
     }
 
     private int compare(Film f0, Film f1) {
-        int result = f0.getLikes().size() - (f1.getLikes().size());
-        return result;
+        return f1.getLikes().size() - (f0.getLikes().size());
     }
 }
